Move withoutNil to objects.ts and add uniqBy util
